zypp-core/parser/json: expose public API for mcp-server-zypp
Move zypp::json::Parser to json/private/jsonparser_p.h (internal only).

Reduce json.h to two public functions: parseDocument() (throws) and
parseDocumentExpected() (zyppng::expected, ZYPP_DLL only).

Move Number/Int/UInt::fromString() to detail free functions in jsonparser_p.h,
removing NG includes from public headers.

Add ZYPP_API to Value, Object, Array, String, Bool, Null, UInt.

Move Json*.h to installed public headers.

Use parseDocumentExpected() in RepoMirrorList.cc to ensure json.o is
referenced and parseDocument() is exported from libzypp.so.
